Make finish_with_err_msg() variadic
i.e. it takes '...' parameters like printf() and passes them to
vprintf() as a list. This means when exiting with an error you no longer
have to make a message with sprintf and can pass a format-string
directly to the function.

Also: guard listfiles allocation of a filename in the array against
surpassing STR_LEN.