listfiles: malloc array in function
Formerly, a stack array would be passed to hold some pointers to
malloc'd strings. This is redundant, inaesthetic and unsafe. Now,
listfiles() counts the files in a directory (no longer the
subdirectories, as I learned how to check if something is a file with
dirent), mallocs an array of char *, and places strdup'd strings into
it.

Add function freefiles() to free such an array.

Use freefiles() where memory formerly leaked.