gen: simplify validition of puzzle
Now, before we even try to check duplicates, we compare the sum of all
cells in the puzzle with the sum of the numbers one to nine, multiplied
by nine: the sum of all cells in a correct Sudoku. If we do not have
that sum, we can return false right away.

Additionally, we now do use our new macros to not check the column, row
and block for each cell, but just each column, row and block once.

The macros for column and row were mixed up, which was fixed.